removed testing routes
