--- views/testingZone.handlebars
@@ -1,1 +1,0 @@
-<script src='/assets/js/testingZone.js'></script>
